move prepends to use the extensions model
