--- app/forms/concerns/flexible_metadata/prepend_admin_set_form.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module FlexibleMetadata
-  module PrependAdminSetForm
-    # override (from Hyrax 2.5.0) - new method to add metadata_context
-    def metadata_context
-      @metadata_context ||= model.metadata_context
-    end
-  end
-end
